chore: small fixes
