+++ Source/PlayerZeroEditor/Public/DeveloperAccounts/Models/ApplicationListResponse.h
@@ -5,7 +5,7 @@
 #include "ApplicationListResponse.generated.h"
 
 USTRUCT(BlueprintType)
-struct PLAYERZEROEDITOR_API FApplication
+struct PLAYERZEROEDITOR_API FPlayerZeroApplication
 {
 	GENERATED_BODY()
 	UPROPERTY(meta = (JsonName = "id"))
@@ -20,5 +20,5 @@ struct PLAYERZEROEDITOR_API FApplicationListResponse : public FApiResponse
 	GENERATED_BODY()
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "API", meta = (JsonName = "data"))
-	TArray<FApplication> Data;
+	TArray<FPlayerZeroApplication> Data;
 };
