+++ Source/PlayerZeroEditor/Public/UI/SPlayerZeroSettingsPanel.h
@@ -15,7 +15,7 @@ class FRpmTextureLoader;
 struct FRpmAsset;
 class FAssetApi;
 struct FAssetListResponse;
-struct FApplication;
+struct FPlayerZeroApplication;
 
 DECLARE_DELEGATE(FOnLogout);
 
@@ -40,14 +40,14 @@ private:
 	TArray<TSharedPtr<FString>> ComboBoxItems;
 	TSharedPtr<STextBlock> SelectedApplicationTextBlock;
 	TArray<UTexture2D*> CharacterStyleTextures;
-	TArray<FApplication> ApplicationList;
+	TArray<FPlayerZeroApplication> ApplicationList;
 	TMap<FString, FCharacterBlueprint> LoadedBlueprints;
 	TArray<TSharedPtr<FPlayerZeroTextureLoader>> ActiveLoaders;
 	FString UserName;
 
 	void LoadBlueprintList();
 	void UpdateErrorMessage(const FString& Message);
-	void PopulateSettingsContent(TArray<FApplication> InApplicationList);
+	void PopulateSettingsContent(TArray<FPlayerZeroApplication> InApplicationList);
 	void SetApplicationIdFromEditorTool(const FString& NewApplicationId);
 	void OnTextureLoaded(UTexture2D* Texture2D, TSharedPtr<SImage> Image, TSharedPtr<FPlayerZeroTextureLoader> RpmTextureLoader);
 	void AddCharacterBlueprint(const FCharacterBlueprint& CharacterBlueprint);
